Move dashboard pages and components
